+++ src/js/constants/index.js
@@ -1,0 +1,2 @@
+export { ERROR_MESSAGE } from "./errorMessage.js";
+export { LOTTO } from "./lotto.js";
